polars_arrow_format

Module ipc

Re-exports

pub use ::planus;

Structs

Binary

Opaque binary data

BinaryBuilder

Builder for serializing an instance of the Binary type.

BinaryRef

Reference to a deserialized Binary.

BinaryView

Logically the same as Binary, but the internal representation uses a view struct that contains the string length and either the string’s entire data inline (for small strings) or an inlined prefix, an index of another buffer, and an offset pointing to a slice in that buffer (for non-small strings).

BinaryViewBuilder

Builder for serializing an instance of the BinaryView type.

BinaryViewRef

Reference to a deserialized BinaryView.

Block

The struct Block in the namespace org.apache.arrow.flatbuf

BlockRef

Reference to a deserialized Block.

BodyCompression

Optional compression for the memory buffers constituting IPC message bodies. Intended for use with RecordBatch but could be used for other message types

BodyCompressionBuilder

Builder for serializing an instance of the BodyCompression type.

BodyCompressionRef

Reference to a deserialized BodyCompression.

Bool

The table Bool in the namespace org.apache.arrow.flatbuf

BoolBuilder

Builder for serializing an instance of the Bool type.

BoolRef

Reference to a deserialized Bool.

Buffer

BufferRef

Reference to a deserialized Buffer.

Date

Date is either a 32-bit or 64-bit signed integer type representing an elapsed time since UNIX epoch (1970-01-01), stored in either of two units:

DateBuilder

Builder for serializing an instance of the Date type.

DateRef

Reference to a deserialized Date.

Decimal

Exact decimal value represented as an integer value in two’s complement. Currently 32-bit (4-byte), 64-bit (8-byte), 128-bit (16-byte) and 256-bit (32-byte) integers are used. The representation uses the endianness indicated in the Schema.

DecimalBuilder

Builder for serializing an instance of the Decimal type.

DecimalRef

Reference to a deserialized Decimal.

DictionaryBatch

For sending dictionary encoding information. Any Field can be dictionary-encoded, but in this case none of its children may be dictionary-encoded. There is one vector / column per dictionary, but that vector / column may be spread across multiple dictionary batches by using the isDelta flag

DictionaryBatchBuilder

Builder for serializing an instance of the DictionaryBatch type.

DictionaryBatchRef

Reference to a deserialized DictionaryBatch.

DictionaryEncoding

The table DictionaryEncoding in the namespace org.apache.arrow.flatbuf

DictionaryEncodingBuilder

Builder for serializing an instance of the DictionaryEncoding type.

DictionaryEncodingRef

Reference to a deserialized DictionaryEncoding.

Duration

The table Duration in the namespace org.apache.arrow.flatbuf

DurationBuilder

Builder for serializing an instance of the Duration type.

DurationRef

Reference to a deserialized Duration.

Field

FieldBuilder

Builder for serializing an instance of the Field type.

FieldNode

FieldNodeRef

Reference to a deserialized FieldNode.

FieldRef

Reference to a deserialized Field.

FixedSizeBinary

The table FixedSizeBinary in the namespace org.apache.arrow.flatbuf

FixedSizeBinaryBuilder

Builder for serializing an instance of the FixedSizeBinary type.

FixedSizeBinaryRef

Reference to a deserialized FixedSizeBinary.

FixedSizeList

The table FixedSizeList in the namespace org.apache.arrow.flatbuf

FixedSizeListBuilder

Builder for serializing an instance of the FixedSizeList type.

FixedSizeListRef

Reference to a deserialized FixedSizeList.

FloatingPoint

The table FloatingPoint in the namespace org.apache.arrow.flatbuf

FloatingPointBuilder

Builder for serializing an instance of the FloatingPoint type.

FloatingPointRef

Reference to a deserialized FloatingPoint.

Footer

FooterBuilder

Builder for serializing an instance of the Footer type.

FooterRef

Reference to a deserialized Footer.

Int

The table Int in the namespace org.apache.arrow.flatbuf

IntBuilder

Builder for serializing an instance of the Int type.

IntRef

Reference to a deserialized Int.

Interval

The table Interval in the namespace org.apache.arrow.flatbuf

IntervalBuilder

Builder for serializing an instance of the Interval type.

IntervalRef

Reference to a deserialized Interval.

KeyValue

KeyValueBuilder

Builder for serializing an instance of the KeyValue type.

KeyValueRef

Reference to a deserialized KeyValue.

LargeBinary

Same as Binary, but with 64-bit offsets, allowing to represent extremely large data values.

LargeBinaryBuilder

Builder for serializing an instance of the LargeBinary type.

LargeBinaryRef

Reference to a deserialized LargeBinary.

LargeList

Same as List, but with 64-bit offsets, allowing to represent extremely large data values.

LargeListBuilder

Builder for serializing an instance of the LargeList type.

LargeListRef

Reference to a deserialized LargeList.

LargeListView

Same as ListView, but with 64-bit offsets and sizes, allowing to represent extremely large data values.

LargeListViewBuilder

Builder for serializing an instance of the LargeListView type.

LargeListViewRef

Reference to a deserialized LargeListView.

LargeUtf8

Same as Utf8, but with 64-bit offsets, allowing to represent extremely large data values.

LargeUtf8Builder

Builder for serializing an instance of the LargeUtf8 type.

LargeUtf8Ref

Reference to a deserialized LargeUtf8.

List

The table List in the namespace org.apache.arrow.flatbuf

ListBuilder

Builder for serializing an instance of the List type.

ListRef

Reference to a deserialized List.

ListView

Represents the same logical types that List can, but contains offsets and sizes allowing for writes in any order and sharing of child values among list values.

ListViewBuilder

Builder for serializing an instance of the ListView type.

ListViewRef

Reference to a deserialized ListView.

Map

A Map is a logical nested type that is represented as

MapBuilder

Builder for serializing an instance of the Map type.

MapRef

Reference to a deserialized Map.

Message

The table Message in the namespace org.apache.arrow.flatbuf

MessageBuilder

Builder for serializing an instance of the Message type.

MessageHeaderBuilder

Builder for serializing an instance of the MessageHeader type.

MessageRef

Reference to a deserialized Message.

Null

These are stored in the flatbuffer in the Type union below

NullBuilder

Builder for serializing an instance of the Null type.

NullRef

Reference to a deserialized Null.

RecordBatch

A data header describing the shared memory layout of a “record” or “row” batch. Some systems call this a “row batch” internally and others a “record batch”.

RecordBatchBuilder

Builder for serializing an instance of the RecordBatch type.

RecordBatchRef

Reference to a deserialized RecordBatch.

RunEndEncoded

Contains two child arrays, run_ends and values. The run_ends child array must be a 16/32/64-bit integer array which encodes the indices at which the run with the value in each corresponding index in the values child array ends. Like list/struct types, the value array can be of any type.

RunEndEncodedBuilder

Builder for serializing an instance of the RunEndEncoded type.

RunEndEncodedRef

Reference to a deserialized RunEndEncoded.

Schema

SchemaBuilder

Builder for serializing an instance of the Schema type.

SchemaRef

Reference to a deserialized Schema.

SparseMatrixIndexCsx

Compressed Sparse format, that is matrix-specific.

SparseMatrixIndexCsxBuilder

Builder for serializing an instance of the SparseMatrixIndexCsx type.

SparseMatrixIndexCsxRef

Reference to a deserialized SparseMatrixIndexCsx.

SparseTensor

The table SparseTensor in the namespace org.apache.arrow.flatbuf

SparseTensorBuilder

Builder for serializing an instance of the SparseTensor type.

SparseTensorIndexBuilder

Builder for serializing an instance of the SparseTensorIndex type.

SparseTensorIndexCoo

SparseTensorIndexCooBuilder

Builder for serializing an instance of the SparseTensorIndexCoo type.

SparseTensorIndexCooRef

Reference to a deserialized SparseTensorIndexCoo.

SparseTensorIndexCsf

Compressed Sparse Fiber (CSF) sparse tensor index.

SparseTensorIndexCsfBuilder

Builder for serializing an instance of the SparseTensorIndexCsf type.

SparseTensorIndexCsfRef

Reference to a deserialized SparseTensorIndexCsf.

SparseTensorRef

Reference to a deserialized SparseTensor.

Struct

A Struct_ in the flatbuffer metadata is the same as an Arrow Struct (according to the physical memory layout). We used Struct_ here as Struct is a reserved word in Flatbuffers

StructBuilder

Builder for serializing an instance of the Struct type.

StructRef

Reference to a deserialized Struct.

Tensor

The table Tensor in the namespace org.apache.arrow.flatbuf

TensorBuilder

Builder for serializing an instance of the Tensor type.

TensorDim

TensorDimBuilder

Builder for serializing an instance of the TensorDim type.

TensorDimRef

Reference to a deserialized TensorDim.

TensorRef

Reference to a deserialized Tensor.

Time

Time is either a 32-bit or 64-bit signed integer type representing an elapsed time since midnight, stored in either of four units: seconds, milliseconds, microseconds or nanoseconds.

TimeBuilder

Builder for serializing an instance of the Time type.

TimeRef

Reference to a deserialized Time.

Timestamp

Timestamp is a 64-bit signed integer representing an elapsed time since a fixed epoch, stored in either of four units: seconds, milliseconds, microseconds or nanoseconds, and is optionally annotated with a timezone.

TimestampBuilder

Builder for serializing an instance of the Timestamp type.

TimestampRef

Reference to a deserialized Timestamp.

TypeBuilder

Builder for serializing an instance of the Type type.

Union

A union is a complex type with children in Field By default ids in the type vector refer to the offsets in the children optionally typeIds provides an indirection between the child offset and the type id for each child typeIds[offset] is the id used in the type vector

UnionBuilder

Builder for serializing an instance of the Union type.

UnionRef

Reference to a deserialized Union.

Utf8

Unicode with UTF-8 encoding

Utf8Builder

Builder for serializing an instance of the Utf8 type.

Utf8Ref

Reference to a deserialized Utf8.

Utf8View

Logically the same as Utf8, but the internal representation uses a view struct that contains the string length and either the string’s entire data inline (for small strings) or an inlined prefix, an index of another buffer, and an offset pointing to a slice in that buffer (for non-small strings).

Utf8ViewBuilder

Builder for serializing an instance of the Utf8View type.

Utf8ViewRef

Reference to a deserialized Utf8View.

Enums

BodyCompressionMethod

Provided for forward compatibility in case we need to support different strategies for compressing the IPC message body (like whole-body compression rather than buffer-level) in the future

CompressionType

The enum CompressionType in the namespace org.apache.arrow.flatbuf

DateUnit

The enum DateUnit in the namespace org.apache.arrow.flatbuf

DictionaryKind

Endianness

Feature

Represents Arrow Features that might not have full support within implementations. This is intended to be used in two scenarios:

IntervalUnit

The enum IntervalUnit in the namespace org.apache.arrow.flatbuf

MessageHeader

MessageHeaderRef

Reference to a deserialized MessageHeader.

MetadataVersion

The enum MetadataVersion in the namespace org.apache.arrow.flatbuf

Precision

The enum Precision in the namespace org.apache.arrow.flatbuf

SparseMatrixCompressedAxis

The enum SparseMatrixCompressedAxis in the namespace org.apache.arrow.flatbuf

SparseTensorIndex

The union SparseTensorIndex in the namespace org.apache.arrow.flatbuf

SparseTensorIndexRef

Reference to a deserialized SparseTensorIndex.

TimeUnit

The enum TimeUnit in the namespace org.apache.arrow.flatbuf

Type

TypeRef

Reference to a deserialized Type.

UnionMode

The enum UnionMode in the namespace org.apache.arrow.flatbuf