polars_arrow_format::ipc

Enum Type

pub enum Type {

    Null(Box<Null>),
    Int(Box<Int>),
    FloatingPoint(Box<FloatingPoint>),
    Binary(Box<Binary>),
    Utf8(Box<Utf8>),
    Bool(Box<Bool>),
    Decimal(Box<Decimal>),
    Date(Box<Date>),
    Time(Box<Time>),
    Timestamp(Box<Timestamp>),
    Interval(Box<Interval>),
    List(Box<List>),
    Struct(Box<Struct>),
    Union(Box<Union>),
    FixedSizeBinary(Box<FixedSizeBinary>),
    FixedSizeList(Box<FixedSizeList>),
    Map(Box<Map>),
    Duration(Box<Duration>),
    LargeBinary(Box<LargeBinary>),
    LargeUtf8(Box<LargeUtf8>),
    LargeList(Box<LargeList>),
    RunEndEncoded(Box<RunEndEncoded>),
    BinaryView(Box<BinaryView>),
    Utf8View(Box<Utf8View>),
    ListView(Box<ListView>),
    LargeListView(Box<LargeListView>),

}

Top-level Type value, enabling extensible type-specific metadata. We can add new logical types to Type without breaking backwards compatibility

Generated from these locations:

Union Type in the file Schema.fbs:442

Variants

Null(Box<Null>)

The variant of type Null in the union Type

Int(Box<Int>)

The variant of type Int in the union Type

FloatingPoint(Box<FloatingPoint>)

The variant of type FloatingPoint in the union Type

Binary(Box<Binary>)

The variant of type Binary in the union Type

Utf8(Box<Utf8>)

The variant of type Utf8 in the union Type

Bool(Box<Bool>)

The variant of type Bool in the union Type

Decimal(Box<Decimal>)

The variant of type Decimal in the union Type

Date(Box<Date>)

The variant of type Date in the union Type

Time(Box<Time>)

The variant of type Time in the union Type

Timestamp(Box<Timestamp>)

The variant of type Timestamp in the union Type

Interval(Box<Interval>)

The variant of type Interval in the union Type

List(Box<List>)

The variant of type List in the union Type

Struct(Box<Struct>)

The variant of type Struct_ in the union Type

Union(Box<Union>)

The variant of type Union in the union Type

FixedSizeBinary(Box<FixedSizeBinary>)

The variant of type FixedSizeBinary in the union Type

FixedSizeList(Box<FixedSizeList>)

The variant of type FixedSizeList in the union Type

Map(Box<Map>)

The variant of type Map in the union Type

Duration(Box<Duration>)

The variant of type Duration in the union Type

LargeBinary(Box<LargeBinary>)

The variant of type LargeBinary in the union Type

LargeUtf8(Box<LargeUtf8>)

The variant of type LargeUtf8 in the union Type

LargeList(Box<LargeList>)

The variant of type LargeList in the union Type

RunEndEncoded(Box<RunEndEncoded>)

The variant of type RunEndEncoded in the union Type

BinaryView(Box<BinaryView>)

The variant of type BinaryView in the union Type

Utf8View(Box<Utf8View>)

The variant of type Utf8View in the union Type

ListView(Box<ListView>)

The variant of type ListView in the union Type

LargeListView(Box<LargeListView>)

The variant of type LargeListView in the union Type

Implementations

impl Type

pub fn builder() -> TypeBuilder<Uninitialized>

Creates a TypeBuilder for serializing an instance of this table.

pub fn create_null( builder: &mut Builder, value: impl WriteAsOffset<Null>, ) -> UnionOffset<Self>

pub fn create_int( builder: &mut Builder, value: impl WriteAsOffset<Int>, ) -> UnionOffset<Self>

pub fn create_floating_point( builder: &mut Builder, value: impl WriteAsOffset<FloatingPoint>, ) -> UnionOffset<Self>

pub fn create_binary( builder: &mut Builder, value: impl WriteAsOffset<Binary>, ) -> UnionOffset<Self>

pub fn create_utf8( builder: &mut Builder, value: impl WriteAsOffset<Utf8>, ) -> UnionOffset<Self>

pub fn create_bool( builder: &mut Builder, value: impl WriteAsOffset<Bool>, ) -> UnionOffset<Self>

pub fn create_decimal( builder: &mut Builder, value: impl WriteAsOffset<Decimal>, ) -> UnionOffset<Self>

pub fn create_date( builder: &mut Builder, value: impl WriteAsOffset<Date>, ) -> UnionOffset<Self>

pub fn create_time( builder: &mut Builder, value: impl WriteAsOffset<Time>, ) -> UnionOffset<Self>

pub fn create_timestamp( builder: &mut Builder, value: impl WriteAsOffset<Timestamp>, ) -> UnionOffset<Self>

pub fn create_interval( builder: &mut Builder, value: impl WriteAsOffset<Interval>, ) -> UnionOffset<Self>

pub fn create_list( builder: &mut Builder, value: impl WriteAsOffset<List>, ) -> UnionOffset<Self>

pub fn create_struct( builder: &mut Builder, value: impl WriteAsOffset<Struct>, ) -> UnionOffset<Self>

pub fn create_union( builder: &mut Builder, value: impl WriteAsOffset<Union>, ) -> UnionOffset<Self>

pub fn create_fixed_size_binary( builder: &mut Builder, value: impl WriteAsOffset<FixedSizeBinary>, ) -> UnionOffset<Self>

pub fn create_fixed_size_list( builder: &mut Builder, value: impl WriteAsOffset<FixedSizeList>, ) -> UnionOffset<Self>

pub fn create_map( builder: &mut Builder, value: impl WriteAsOffset<Map>, ) -> UnionOffset<Self>

pub fn create_duration( builder: &mut Builder, value: impl WriteAsOffset<Duration>, ) -> UnionOffset<Self>

pub fn create_large_binary( builder: &mut Builder, value: impl WriteAsOffset<LargeBinary>, ) -> UnionOffset<Self>

pub fn create_large_utf8( builder: &mut Builder, value: impl WriteAsOffset<LargeUtf8>, ) -> UnionOffset<Self>

pub fn create_large_list( builder: &mut Builder, value: impl WriteAsOffset<LargeList>, ) -> UnionOffset<Self>

pub fn create_run_end_encoded( builder: &mut Builder, value: impl WriteAsOffset<RunEndEncoded>, ) -> UnionOffset<Self>

pub fn create_binary_view( builder: &mut Builder, value: impl WriteAsOffset<BinaryView>, ) -> UnionOffset<Self>

pub fn create_utf8_view( builder: &mut Builder, value: impl WriteAsOffset<Utf8View>, ) -> UnionOffset<Self>

pub fn create_list_view( builder: &mut Builder, value: impl WriteAsOffset<ListView>, ) -> UnionOffset<Self>

pub fn create_large_list_view( builder: &mut Builder, value: impl WriteAsOffset<LargeListView>, ) -> UnionOffset<Self>