polars_arrow_format::ipc

Struct DictionaryBatch

pub struct DictionaryBatch {
    pub id: i64,
    pub data: Option<Box<RecordBatch>>,
    pub is_delta: bool,
}

For sending dictionary encoding information. Any Field can be dictionary-encoded, but in this case none of its children may be dictionary-encoded. There is one vector / column per dictionary, but that vector / column may be spread across multiple dictionary batches by using the isDelta flag

Generated from these locations:

Table DictionaryBatch in the file Message.fbs:127

Fields
id: i64

The field id in the table DictionaryBatch

data: Option<Box<RecordBatch>>

The field data in the table DictionaryBatch

is_delta: bool

If isDelta is true the values in the dictionary are to be appended to a dictionary with the indicated id. If isDelta is false this dictionary should replace the existing dictionary.

Implementations

impl DictionaryBatch

pub fn builder() -> DictionaryBatchBuilder<()>

Creates a DictionaryBatchBuilder for serializing an instance of this table.

pub fn create( builder: &mut Builder, field_id: impl WriteAsDefault<i64, i64>, field_data: impl WriteAsOptional<Offset<RecordBatch>>, field_is_delta: impl WriteAsDefault<bool, bool>, ) -> Offset<Self>